+++ packages/core/src/extensions/atom-mark-navigation.ts
@@ -16,9 +16,9 @@ import { Decoration, DecorationSet } from '@prosekit/pm/view'
 import { getIsComposing } from '../utils/composition.ts'
 
 import { hasPointerSelectionTransaction } from '../utils/transaction.ts'
-import { getMarkRangeAt } from './get-mark-range-at.ts'
 import { getMarkMode, type MarkMode } from './mark-mode.ts'
 import type { MarkName } from './mark-names.ts'
+import { getMarkRangeAt } from './mark-range.ts'
 
 /**
  * The source marks whose mark views hide the raw text behind a rendered
