+++ packages/core/src/extensions/caret-rect.ts
@@ -3,8 +3,8 @@ import type { EditorView } from '@prosekit/pm/view'
 import { tryCoordsAtPos, type CaretCoords } from '../utils/caret-coords.ts'
 
 import { ATOM_SOURCE_MARK_NAMES } from './atom-mark-navigation.ts'
-import { getMarkRangeAt } from './get-mark-range-at.ts'
 import { getHiddenRunAfter, getHiddenRunBefore } from './hidden-run.ts'
+import { getMarkRangeAt } from './mark-range.ts'
 
 export interface CaretRect {
   left: number
