+++ packages/core/src/extensions/tag-click.ts
@@ -1,7 +1,7 @@
 import type { PlainExtension } from '@prosekit/core'
 import { PluginKey, type EditorState } from '@prosekit/pm/state'
 
-import { getMarkRangeAt } from './get-mark-range-at.ts'
+import { getMarkRangeAt } from './mark-range.ts'
 import { defineMarkClickHandler } from './mark-click.ts'
 
 const tagClickKey = new PluginKey('meowdown-tag-click')
